Reorgranize handlers
